requirements.txt
apispec==4.5.0
click==7.1.2
Flask==1.1.2
flask-apispec==0.11.0
Flask-Cors==3.0.10
Flask-JWT-Extended==4.2.0
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
importlib-metadata==4.0.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
marshmallow==3.12.1
passlib==1.7.4
PyJWT==2.1.0
python-dotenv==0.17.1
six==1.15.0
SQLAlchemy==1.4.13
typing-extensions==3.10.0.0
webargs==8.0.0
Werkzeug==1.0.1
zipp==3.4.1
